PKI.digest {PKI}	R Documentation

Compute digest sum based on SHA1, SHA256 or MD5 hash functions

Description

PKI.digest computes digsest sum based on the hash function specified

Usage

PKI.digest(what, hash = c("SHA1", "SHA256", "MD5"))

Arguments

what
raw vector of bytes to digest

hash
type of the hash function. Note that "MD5" should not be used for cryptographic purposes as it is not secure

Value

Raw vector containg the hash

Author(s)

Simon Urbanek

See Also

PKI.sign

Examples

  PKI.digest(as.raw(1:10))

[Package PKI version 0.1-5.1 Index]