PKI.random {PKI}	R Documentation

Generate cryptographically strong pseudo-random bytes.

Description

PKI.random generates n cryptographically strong pseudo-random bytes.

Usage

PKI.random(n)

Arguments

n
non-negative integer, number of bytes to generate

Details

PKI.random is the preferred way to generate cryptographically strong random content that can be used as keys, seeds etc. Not to be confused with random number generators in R, it is entirely separate for cryptographics purposes.

Value

Raw vector of n cryptographically strong pseudo-random bytes.

Author(s)

Simon Urbanek

Examples

  PKI.random(10)

[Package PKI version 0.1-5.1 Index]