X509 {PKI}	R Documentation

Public Key Instraftructure (X509) functions

Description

PKI.load.cert creates a certificate obejct from a string, connection or file.

PKI.verifyCA verifies a certificate against a given chain of trust.

PKI.pubkey extracts public key from a certificate.

PKI.get.subject extracts the subject name from the certificate.

Usage

PKI.load.cert(what, format = c("PEM", "DER"), file)
PKI.verifyCA(certificate, ca)
PKI.pubkey(certificate)
PKI.get.subject(certificate)

Arguments

what
string, raw vector or connection to load the certificate from

format
format used to encode the certificate

file
filename to load the certificate from - what and file are mutually exclusive

certificate
a certificate object (as returned by PKI.load.cert)

ca
a certificate object of the Certificate Authority (CA) or a list of such objects if multiple CAs are involved

Value

PKI.load.code: a certificate object

PKI.verifyCA: TRUE is the certificate can be trusted, FALSE otherwise

PKI.pubkey: public key object

PKI.get.subject: string containing the subject information in one-line RFC2253 format but in UTF8 encoding instead of MBS escapes. NOTE: this is experimantal, we may choose to parse the contents and return it in native R form as a named vector instead.

Author(s)

Simon Urbanek

Examples

  ca <- PKI.load.cert(file=system.file("certs", "RForge-ca.crt", package="PKI"))
  my.cert <- PKI.load.cert(readLines(system.file("certs", "demo.crt", package="PKI")))
  PKI.verifyCA(my.cert, ca)
  PKI.pubkey(my.cert)
  PKI.get.subject(my.cert)

[Package PKI version 0.1-5.1 Index]